pyproject.toml
[tool.poetry]
name = "secure_coding_platform"
version = "0.1.0"
description = "SCCAP — Secure Coding & Compliance Automation Platform. Open-source AI-powered code auditing, remediation, and compliance tooling."
authors = ["dev@nerdykrishna.com"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
# This project requires Python 3.12.x
# Core deps: everything the API container needs at runtime. The heavy
# ML / AST-parsing stack lives in the optional "worker" group below and
# is only installed in the worker image.
aio-pika = ">=9.5.5,<10.0.0"
alembic = ">=1.16.1,<2.0.0"
anthropic = ">=0.96.0,<1.0.0"
asyncpg = ">=0.30.0,<0.31.0"
cryptography = "^46.0.7"
email-validator = ">=2.2.0,<3.0.0" # Useful for FastAPI Users
# Range widened from `<0.116.0` to allow starlette ≥ 0.47.2 (CVE
# fixes — GHSA-2c2j-9gv5-cj73 / GHSA-7f5h-v6xp-fcq8). FastAPI 0.115
# pinned starlette `<0.47.0`; 0.116+ relaxes that.
fastapi = ">=0.115.9,<0.121.0"
fastapi-mail = ">=1.5.9,<1.6.0"
# Bumped 14.x → 15.x for GHSA-5j53-63w8-8625 (the 14.x line pins
# pyjwt = 2.10.1 and python-multipart = 0.0.20 as exact versions,
# blocking three more transitive CVE fixes; 15.x relaxes both).
fastapi-users = {extras = ["sqlalchemy"], version = "^15.0.2"}
gitpython = "^3.1.44"
google-api-core = "^2.24.2"
google-genai = "^1.16.1"
itsdangerous = ">=2.2.0,<3.0.0"
# Phase I.1 — LangChain 1.x line. langchain-core 1.3 is the first release
# of the 1.x series we align to; picks up: standardized `response.usage_metadata`
# across providers, LangGraph 1.x compatibility, native support for Gemini 3.x
# via langchain-google-genai 4.2.2, and structured-output strategies that
# removed the need for our provider-specific fallbacks.
langchain-anthropic = "^1.4.1"
langchain-core = "^1.3.0"
langchain-google-genai = "^4.2.2"
langchain-openai = "^1.2.0"
langgraph = "^1.1.9"
langgraph-checkpoint-postgres = "^3.0.5"
# Phase I.2 — LiteLLM as the token-counter + price-table source of truth.
# Unified tokenizers across providers; community-maintained pricing map at
# litellm.model_prices_and_context_window. Pinned to ^1.76 because 1.80+
# exact-pins openai/tiktoken/python-dotenv in ways that clash with
# langchain-openai 1.2.x; the 1.76 line still uses range constraints
# (openai>=1.99.5, python-dotenv>=0.2.0, tiktoken>=0.7).
litellm = "^1.76.0"
# Phase I.3 — Pydantic AI for node-level structured output with typed
# validation-retry. No extras — the provider SDKs (openai, anthropic,
# google-genai) are already present via LangChain; we only need the
# core Agent class + pydantic_ai.models.{openai,anthropic,google} classes.
pydantic-ai = "^1.86.1"
# Phase I.4 — FastMCP server exposing scan/advisor/remediation tools to
# external agents (Claude Code, Cursor, other MCP clients). Mounted on
# the same FastAPI app; shares JWT auth with the REST API via a
# custom TokenVerifier that reuses CustomCookieJWTStrategy.
fastmcp = "^3.2.4"
# Optional LLM observability via self-hosted Langfuse v3. SDK is OTel-
# native; no separate `opentelemetry-*` deps are needed because the
# Langfuse SDK ships its own OTel wiring. Disabled at runtime by
# default (LANGFUSE_ENABLED=false). See `app.infrastructure.observability`.
langfuse = "^3"
networkx = "^3.5"
pandas = "^2.3.0"
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
# Vector store + embedder. fastembed (Qdrant-maintained, MIT) provides
# the same all-MiniLM-L6-v2 ONNX bundle the prior chromadb-bundled
# embedder used; vectors are byte-equivalent (max per-dim diff ~6e-9).
qdrant-client = "~1.17"
fastembed = "~0.8"
pydantic = ">=2.12,<3.0.0"  # pydantic-ai 1.86 requires >=2.12
pydantic-settings = "^2.9.1"
python = "~3.12"
python-dotenv = ">=1.1.0,<2.0.0"
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.41,<3.0.0"}
tiktoken = "^0.12.0"  # bumped for pydantic-ai-slim[openai] compat
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.40.0"}
# Bumped from `^0.0.20` to pick up GHSA-wp53-j4wj-2cfg (0.0.22) and
# GHSA-mj87-hwqh-73pj (0.0.26). fastapi-users 15.0.4+ requires the
# same range; alignment avoids a resolver split.
python-multipart = ">=0.0.22,<0.1.0"
cvss = "^3.4"
# bandit ships in both API + worker images so the deterministic SAST
# pre-pass node (`infrastructure/scanners/bandit_runner.py`) can invoke
# `/app/.venv/bin/bandit` as a subprocess in the worker.
bandit = "^1.7"
croniter = "^5.0"
# Enterprise SSO. httpx-oauth provides the OIDC client used in
# infrastructure/auth/sso/oidc.py (PKCE + state + nonce). pyjwt[crypto]
# is needed to validate the IdP-signed id_token against the JWKS
# (httpx-oauth returns the raw token but does NOT verify the signature
# — see threat-model M1).
# python3-saml is OneLogin's SAML 2.0 SP library; it requires libxmlsec1
# system packages (added to Dockerfile in the same change).
httpx-oauth = ">=0.16,<1.0"
pyjwt = {extras = ["crypto"], version = ">=2.10,<3.0"}
python3-saml = ">=1.16,<2.0"


# Worker-only deps: AST parsing (tree-sitter). The embedder was
# migrated off sentence-transformers / torch in Phase H.1.1, then
# repointed at fastembed (bundled MiniLM-L6-v2 ONNX) when ADR-008
# replaced ChromaDB with Qdrant — vectors are byte-equivalent, so
# the worker still doesn't need those heavy packages. Keeping the
# group structure for future worker-only additions.
[tool.poetry.group.worker]
optional = true

[tool.poetry.group.worker.dependencies]
tree-sitter = ">=0.21.0,<0.22.0"
tree-sitter-languages = "^1.10.2"
# Semgrep is installed in an isolated /opt/semgrep-venv inside the
# worker Dockerfile (see Dockerfile worker stage). It's NOT in the main
# Poetry dep tree because semgrep pins `rich<13.6` while fastmcp ^3.2.4
# requires `rich>=13.9.4` — the conflict is unresolvable. The runner
# resolves the binary via `app.infrastructure.scanners.bandit_runner._resolve_binary`
# which honors `SEMGREP_BINARY` env var → `shutil.which("semgrep")` →
# fallback path `/opt/semgrep-venv/bin/semgrep`.


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
black = "^25.1.0"
# black = "^24.4.2"
mypy = "^1.10.0"
# bandit lives in main deps because it's invoked at runtime by the SAST
# pre-pass node; pip-audit + pre-commit stay in dev.
pip-audit = "^2.7"
pre-commit = "^3.7"

# Test deps live in their own optional group so runtime images don't ship
# them. Install with `poetry install --with test` locally or in CI.
[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.23.7"
pytest-cov = "^7.0.0"
httpx = ">=0.28,<1.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
# The Playwright UI setup test needs a running UI container and a browser;
# keep it manual and off the default pytest target. Run explicitly with
# `pytest tests/test_ui_setup.py` when exercising the setup flow.
addopts = "--ignore=tests/test_ui_setup.py"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
ignore = ["E501"] # Ignore line length violations as we let black/ruff format handle it best effort

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
